placed PathToDirectory function in util.cc
