+++ cpp/src/graphar/graph_info.cc
@@ -1467,26 +1467,4 @@ Status GraphInfo::Save(const std::string& path) const {
   GAR_ASSIGN_OR_RAISE(auto yaml_content, this->Dump());
   return fs->WriteValueToFile(yaml_content, no_url_path);
 }
-
-namespace util {
-std::string PathToDirectory(const std::string& path) {
-  if (path.rfind("s3://", 0) == 0) {
-    size_t t = path.find_last_of('?');
-    // Your fix here
-    std::string prefix = (t == std::string::npos) ? path : path.substr(0, t);
-    std::string suffix = (t == std::string::npos) ? "" : path.substr(t);
-
-    const size_t last_slash_idx = prefix.rfind('/');
-    if (std::string::npos != last_slash_idx) {
-      return prefix.substr(0, last_slash_idx + 1) + suffix;
-    }
-  } else {
-    const size_t last_slash_idx = path.rfind('/');
-    if (std::string::npos != last_slash_idx) {
-      return path.substr(0, last_slash_idx + 1);
-    }
-  }
-  return path;
-}
-}  // namespace util
 }  // namespace graphar
