+++ cpp/src/graphar/util.cc
@@ -106,4 +106,23 @@ std::string ValueGetter<std::string>::Value(const void* data, int64_t offset) {
       reinterpret_cast<const arrow::LargeStringArray*>(data)->GetView(offset));
 }
 
+std::string PathToDirectory(const std::string& path) {
+  if (path.rfind("s3://", 0) == 0) {
+    size_t t = path.find_last_of('?');
+
+    std::string prefix = (t == std::string::npos) ? path : path.substr(0, t);
+    std::string suffix = (t == std::string::npos) ? "" : path.substr(t);
+
+    const size_t last_slash_idx = prefix.rfind('/');
+    if (std::string::npos != last_slash_idx) {
+      return prefix.substr(0, last_slash_idx + 1) + suffix;
+    }
+  } else {
+    const size_t last_slash_idx = path.rfind('/');
+    if (std::string::npos != last_slash_idx) {
+      return path.substr(0, last_slash_idx + 1);
+    }
+  }
+  return path;
+}
 }  // namespace graphar::util
